Add modern video player example with full API coverage
New Flutter example app at example/player/ demonstrating all
video_player API options: play/pause, seek with scrubber, volume
control with mute toggle, playback speed selection, and loop toggle.

Includes an editable media library pre-populated with Blender
Foundation open-source films (Big Buck Bunny, Sintel, Elephants Dream)
and local asset support. Users can add, edit, and delete library items
with configurable name, URL, and source type (network/local).

Key implementation details:
- Material 3 dark theme with NavigationRail layout
- Pipeline auto-started after creation to unblock initialization
- Proper async disposal when switching between videos
- Graceful handling of playback completion without seek-to-end
- MiniController extended with setVolume() and setLooping() methods
