--- packages/video_player/video_player_linux/example/README.md
@@ -1,9 +1,0 @@
-# Platform Implementation Test App
-
-This is a test app for manual testing and automated integration testing
-of this platform implementation. It is not intended to demonstrate actual use of
-this package, since the intent is that plugin clients use the app-facing
-package.
-
-Unless you are making changes to this implementation package, this example is
-very unlikely to be relevant.
