+++ packages/video_player/video_player_linux/example/player/.desktop-homescreen/data/flutter_assets
@@ -1,0 +1,1 @@
+../../build/flutter_assets
