--- packages/video_player/video_player_linux/example/pubspec.yaml
@@ -1,33 +1,0 @@
-name: video_player_example
-description: Demonstrates how to use the video_player plugin.
-publish_to: none
-
-environment:
-  sdk: ">=3.0.0 <4.0.0"
-  flutter: ">=3.10.0"
-
-dependencies:
-  flutter:
-    sdk: flutter
-  video_player_platform_interface: ">=6.1.0 <7.0.0"
-  video_player_linux:
-    # When depending on this package from a real application you should use:
-    #   video_player_linux: ^x.y.z
-    # See https://dart.dev/tools/pub/dependencies#version-constraints
-    # The example app is bundled with the plugin so we use a path dependency on
-    # the parent directory to use the current plugin's version.
-    path: ../
-
-dev_dependencies:
-  flutter_test:
-    sdk: flutter
-  integration_test:
-    sdk: flutter
-  path_provider: ^2.0.6
-  test: any
-
-flutter:
-  uses-material-design: true
-  assets:
-    - assets/flutter-mark-square-64.png
-    - assets/Butterfly-209.mp4
