--- packages/video_player/video_player_linux/example/test_driver/integration_test.dart
@@ -1,7 +1,0 @@
-// Copyright 2013 The Flutter Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-import 'package:integration_test/integration_test_driver.dart';
-
-Future<void> main() => integrationDriver();
